requirements.txt
annotated-types==0.6.0
anyio==3.7.1
boto3==1.33.1
botocore==1.33.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
mangum==0.17.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.8.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
torch==2.1.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
